test cases added for JCasC
