+++ src/test/resources/org/jenkinsci/plugins/ewm/casc/obsolete-configuration-as-code.yaml
@@ -6,7 +6,7 @@ jenkins:
 configuration-as-code:
   deprecated: warn
 unclassified:
-  exwsallocatestep:
+  exwsAllocateStep:
     diskPools:
       - diskPoolId: "diskpool1"
         disks:
